*** empty log message ***

+++ QCDAnalysis/interface/QCDEventHdr.h
@@ -21,7 +21,7 @@ class QCDEventHdr
       void setBS(float fBSx, float fBSy, float fBSz) {mBSx = fBSx; mBSy = fBSy; mBSz = fBSz;}
       void setHCALNoise(bool fNoise) {mHCALNoise = fNoise;}
       void setPU(int fNBX, int fOOTPUEarly, int fOOTPULate, int fINTPU) {mNBX = fNBX; mOOTPUEarly = fOOTPUEarly; mOOTPULate = fOOTPULate; mINTPU = fINTPU;}
-      void setTrPu(int fTrPu) {mTrPu = fTrPu;} // setting the true PU 
+      void setTrPu(float fTrPu) {mTrPu = fTrPu;} // setting the true PU 
       //------------ Get methods ------------------------------
       int runNo()           const {return mRun;} 
       int event()           const {return mEvent;} 
@@ -34,7 +34,7 @@ class QCDEventHdr
       int intpu()           const {return mINTPU;}
       int nbx()             const {return mNBX;} 
       int pu()              const {return mOOTPUEarly+mOOTPULate+mINTPU;}
-      int trpu()            const {return mTrPu;} // get method for True number of interaction
+      float trpu()          const {return mTrPu;} // get method for True number of interaction
       bool isPVgood()       const {return mIsPVgood;}
       bool hcalNoise()      const {return mHCALNoise;}
       float PVndof()        const {return mPVndof;} 
@@ -76,7 +76,7 @@ class QCDEventHdr
         //---- ndof for the primary vertex ------------
         float mPVndof;
         //---- event weight for pileup reweighting --------//
-        int mTrPu;
+        float mTrPu;
         //---- position of the primary vertex --------- 
         float mPVx;
         float mPVy;
